Bowman Camp Stove Knob
Modified version of https://www.thingiverse.com/thing:2224553 , logo replaced with the letters BOW for Camp Bowman.

This file can be opened with Freecad and further modified if needed.  Freecan can also be used to export an STL file which then can be run through a slicer to be 3d printed.